refactor(deeptrace): modularize library
